Create health router
